// src/app/models/subsystem-model.ts
export interface SubsystemModel {

  subCategoryId: number;
  subcategoryName: string;
  subcategoryStatus: number;

  subcategoryCreatedBy?: number;
  subcategoryCreatedTime?: string;

  subcategoryModifiedBy?: number | null;
  subcategoryModifiedTime?: string | null;

  // subCategoryId: number;
  // name: string;
  // status: number;

  // subcategoryCreatedBy: number;
  // subcategoryCreatedTime: string;

  // subcategoryModifiedBy?: number | null;
  // subcategoryModifiedTime?: string | null;
}